requirements.txt
docxtpl >= 0.6.3
django >= 2.2.5
django-filter >= 1.0.1
djangorestframework >= 3.10.3
djangorestframework_simplejwt >= 4.3.0
jinja2 >= 2.10.3
markdown >= 3.0.0
factory_boy >= 2.12.0
psycopg2 >= 2.5.4
dateparser >= 0.7.2
